# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateUser": {
        "additionalProperties": false,
        "example": {
          "email": "singer@example.com",
          "role": "default"
        },
        "properties": {
          "default_collection": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "required": [
          "email"
        ],
        "type": "object"
      },
...
      "Role": {
        "enum": [
          "default",
          "admin"
        ],
        "type": "string"
      },
...
      "User": {
        "example": {
          "avatar_blob_id": null,
          "created_at": "2026-01-01T12:00:00Z",
          "default_collection": null,
          "email": "singer@example.com",
          "id": "usr_example",
          "oauth_avatar_blob_id": null,
          "oauth_picture_url": null,
          "role": "default"
        },
        "properties": {
          "avatar_blob_id": {
            "description": "User-uploaded profile image; takes precedence over [`Self::oauth_avatar_blob_id`].",
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "default_collection": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "oauth_avatar_blob_id": {
            "description": "Backend-cached OAuth profile image (`GET /api/v1/blobs/{id}/data`).",
            "type": [
              "string",
              "null"
            ]
          },
          "oauth_picture_url": {
            "description": "Last `picture` claim URL seen from OIDC (used to detect when to re-fetch the cached avatar).",
            "type": [
              "string",
              "null"
            ]
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "created_at"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "SessionCookie": {
        "description": "Session cookie returned after a successful authentication flow",
        "in": "cookie",
        "name": "sso_session",
        "type": "apiKey"
      },
      "SessionToken": {
        "description": "Session override using `Authorization: Bearer <session>` header",
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      }
    }
  },
  "info": {
...
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "title": "Worship Viewer API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "operationId": "get_users",
        "parameters": [
          {
            "description": "Page index, zero-based. Defaults to 0. See `docs/business-logic-constraints/list-pagination.md` (Track A: `X-Total-Count` is pre-pagination total; last page when `items.len() < page_size` or empty).",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "type": [
                "integer",
                "null"
              ]
            }
          },
          {
            "description": "Items per page. Must be 1–500. Defaults to 50.",
            "example": 50,
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "format": "int32",
              "maximum": 500,
              "minimum": 1,
              "type": [
                "integer",
                "null"
              ]
            }
          },
          {
            "description": "Optional case-insensitive substring filter on email or user id. Whitespace-only is treated as absent.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/User"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Returns list of all users. `X-Total-Count` header contains the total matching user count (before pagination)."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid pagination parameters"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Admin role required"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "API rate limit exceeded; see `Retry-After` and `X-RateLimit-*` response headers"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Failed to list users"
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "SessionToken": []
          }
        ],
        "tags": [
          "Users"
        ]
      },
      "post": {
        "operationId": "create_user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUser"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "description": "Creates a new user"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid request payload"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Admin role required"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "User with that email already exists"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "API rate limit exceeded; see `Retry-After` and `X-RateLimit-*` response headers"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Failed to create user"
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "SessionToken": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/me": {
      "get": {
        "operationId": "get_users_me",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "description": "Returns the currently authenticated user"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Authentication required"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "API rate limit exceeded; see `Retry-After` and `X-RateLimit-*` response headers"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Failed to load user session"
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "SessionToken": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/v1/users/me/profile-picture": {
      "delete": {
        "operationId": "delete_profile_picture",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "description": "Removed uploaded avatar if any; returns updated `User`"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Authentication required"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "API rate limit exceeded"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Server error"
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "SessionToken": []
          }
        ],
        "tags": [
          "Users"
        ]
      },
      "put": {
        "operationId": "put_profile_picture",
        "requestBody": {
          "content": {
            "image/jpeg": {
              "schema": {
                "items": {
                  "format": "int32",
                  "minimum": 0,
                  "type": "integer"
                },
                "type": "array"
              }
            }
          },
          "description": "Raw JPEG or PNG bytes",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "description": "Uploaded; returns updated `User`"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid Content-Type or image"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Authentication required"
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Payload too large"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "API rate limit exceeded"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Server error"
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "SessionToken": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/me/session": {
      "get": {
        "operationId": "get_current_session_for_user",
        "parameters": [
          {
            "description": "Optional `user` to embed full user (default: `id`+`email` link).",
            "in": "query",
            "name": "expand",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionBody"
                }
              }
            },
            "description": "Returns the session credential used for this request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Authentication required"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Session not found"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "API rate limit exceeded; see `Retry-After` and `X-RateLimit-*` response headers"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Failed to fetch session"
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "SessionToken": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/v1/users/me/sessions": {
      "get": {
        "operationId": "get_sessions_for_current_user",
        "parameters": [
          {
            "description": "Zero-based page; defaults to 0. `X-Total-Count` is the total before paging (`list-pagination.md`).",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "type": [
                "integer",
                "null"
              ]
            }
          },
          {
            "description": "Items per page. Must be 1–500. Defaults to 50 when omitted.",
            "example": 50,
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "format": "int32",
              "maximum": 500,
              "minimum": 1,
              "type": [
                "integer",
                "null"
              ]
            }
          },
          {
            "description": "Optional case-insensitive substring on session id, user id, or user email. Whitespace-only is treated as absent.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional comma-separated relations (`user` = embed full user on each session; default is `id`+`email` link only).",
            "in": "query",
            "name": "expand",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionBody"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Returns active sessions for the current user. `X-Total-Count` is the total before paging."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid pagination parameters"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Authentication required"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "API rate limit exceeded; see `Retry-After` and `X-RateLimit-*` response headers"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Failed to list sessions for current user"
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "SessionToken": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Same origin as the web app (override per deployment).",
      "url": "/"
    },
    {
      "description": "Production deployment (public web app).",
      "url": "https://app.worshipviewer.com"
    }
  ],
...
}